+++ backend/src/main/java/com/eshop/backend/product/catalog/service/ProductServiceImp.java
@@ -51,7 +51,7 @@ public class ProductServiceImp implements ProductService {
     }
 
     @Override
-    public Product getById(int id) {
+    public Product getById(Long id) {
         return productdao.getById(id);
     }
